# src/main/java/com/speed_liv/menu/api/MenusApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.speed_liv.menu.api;

import com.speed_liv.menu.model.Plat;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-14T09:37:49.653006400+01:00[Africa/Casablanca]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "menus", description = "the menus API")
public interface MenusApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /menus : Liste de tous les plats disponibles
     *
     * @return Liste des plats (status code 200)
     */
    @Operation(
        operationId = "menusGet",
        summary = "Liste de tous les plats disponibles",
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste des plats", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Plat.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menus",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Plat>> menusGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"prix\" : 0.8008282, \"id\" : \"id\", \"nom\" : \"nom\", \"disponible\" : true }, { \"prix\" : 0.8008282, \"id\" : \"id\", \"nom\" : \"nom\", \"disponible\" : true } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /menus/{id} : Détails d’un plat
     *
     * @param id  (required)
     * @return Détail d’un plat (status code 200)
     */
    @Operation(
        operationId = "menusIdGet",
        summary = "Détails d’un plat",
        responses = {
            @ApiResponse(responseCode = "200", description = "Détail d’un plat", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Plat.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menus/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Plat> menusIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"prix\" : 0.8008282, \"id\" : \"id\", \"nom\" : \"nom\", \"disponible\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}